seprate transaction_recorder.rs
