+++ core/src/validator.rs
@@ -76,8 +76,9 @@ use {
         datapoint_info, metrics::metrics_config_sanity_check, poh_timing_point::PohTimingSender,
     },
     solana_poh::{
-        poh_recorder::{PohRecorder, TransactionRecorder},
+        poh_recorder::PohRecorder,
         poh_service::{self, PohService},
+        transaction_recorder::TransactionRecorder,
     },
     solana_rayon_threadlimit::{get_max_thread_count, get_thread_count},
     solana_rpc::{
