+++ poh/src/lib.rs
@@ -2,6 +2,7 @@
 pub mod leader_bank_notifier;
 pub mod poh_recorder;
 pub mod poh_service;
+pub mod transaction_recorder;
 
 #[macro_use]
 extern crate solana_metrics;
